requirements.txt
linkup-sdk
python-dotenv
typer
click
flask
fastmcp
mcp
personalization-client@https://fastino-ai.github.io/Pioneer/simple/
